Cleanup STAR test suite.
